ir: Anticipate clustering toposort
